+++ src/all/background_page/event/debugEvents.js
@@ -4,6 +4,7 @@
  * @copyright (c) 2017 Passbolt SARL
  * @licence GNU Affero General Public License http://www.gnu.org/licenses/agpl-3.0.en.html
  */
+var Log = require('../model/log').Log;
 var tabsController = require('../controller/tabsController');
 
 var listen = function (worker) {
@@ -17,6 +18,17 @@ var listen = function (worker) {
    */
   worker.port.on('passbolt.debug.open-tab', function (url) {
     tabsController.open(url);
+  });
+
+  /*
+   * Log.
+   *
+   * @listens passbolt.debug.log
+   * @param requestId {uuid} The request identifier
+   * @param data {mixed} The data to log.
+   */
+  worker.port.on('passbolt.debug.log', function (data) {
+    Log.write({level: 'error', message: data});
   });
 
 };